Very simple web interface
This change serves a JSON API on port $ROCKET_PORT that returns the
current count of how many tweets have been made about each keyword of
each sentiment.

This implementation uses a mutex-guarded String of the JSON to be
returned. Whenever the ingester thread ingests a new tweet, every 13th
time, it will try_lock on this mutex and try to update it.

The HTTP GET, on the other hand, will block until it can lock this mutex
and return the string to the user. However, the HTTP GET never waits for
the twitter stream to advance - it will simply return a very-slightly
out of date sentiment summary.