Minimal working thing
I hesitate to say "viable product" - but this does technically call the
twitter API, and then serve up an API that summarizes tweet sentiment,
and then call _that_ API and plot it in rust/wasm. So it feels like
progress.

Next steps are to clean up the plot so that anyone would ever want to
look at it.