Working single binary
This commit marks a single working binary that can be rebuilt and run
from start-server.sh. It will serve all its own HTML and JavaScript.
When launched, the process will open a connection to the twitter
streaming API, update an internal structure representing a crude
sentiment analysis of that API, and export a route that returns the
current state of that data structure.