Periodically clear sentiment cache
Because the front-end starts to look terrible at high numbers of Tweets,
when there are too many tweets just drop the analysis and start over.
This is a terrible hack that would be better addressed by updating the
front-end to scale when it has large numbers in it, but I don't have
time for that.